added unit tests for converter of current and forecast weather (TDD)

+++ OpenWeather-CSharp/Domain/OpenWeatherMap/Model/Sys.cs
@@ -8,7 +8,7 @@ namespace Domain.OpenWeatherMap.Model
 
         public int Type { get; set; }
 
-        public int Message { get; set; }
+        public double Message { get; set; }
 
         public string Country { get; set; }
