+++ OpenWeather-CSharp/Services/OpenWeatherMap/Dto/SysDto.cs
@@ -8,7 +8,7 @@ namespace Services.OpenWeatherMap.Dto
 
         public int Type { get; set; }
 
-        public int Message { get; set; }
+        public double Message { get; set; }
 
         public string Country { get; set; }
