+++ OpenWeather-CSharp/Services/OpenWeatherMap/Dto/WeatherForecastDto.cs
@@ -7,7 +7,7 @@ namespace Services.OpenWeatherMap.Dto
     {
         public string Cod { get; set; }
 
-        public int Message { get; set; }
+        public double Message { get; set; }
 
         public uint Count { get; set; }
