+++ OpenWeather-CSharp/Web/Models/OpenWeatherMap/WeatherForecastViewModel.cs
@@ -7,7 +7,7 @@ namespace Web.Models.OpenWeatherMap
     {
         public string Cod { get; set; }
 
-        public int Message { get; set; }
+        public double Message { get; set; }
 
         public uint Count { get; set; }
